organise babbage tests
