+++ cardano-chain-gen/cardano-chain-gen.cabal
@@ -159,6 +159,14 @@ test-suite cardano-chain-gen
                         Test.Cardano.Db.Mock.Property.Property
                         Test.Cardano.Db.Mock.Unit.Alonzo
                         Test.Cardano.Db.Mock.Unit.Babbage
+                        Test.Cardano.Db.Mock.Unit.Babbage.Other
+                        Test.Cardano.Db.Mock.Unit.Babbage.InlineAndReference
+                        Test.Cardano.Db.Mock.Unit.Babbage.Plutus
+                        Test.Cardano.Db.Mock.Unit.Babbage.Reward
+                        Test.Cardano.Db.Mock.Unit.Babbage.Rollback
+                        Test.Cardano.Db.Mock.Unit.Babbage.Simple
+                        Test.Cardano.Db.Mock.Unit.Babbage.Stake
+                        Test.Cardano.Db.Mock.Unit.Babbage.Tx
                         Test.Cardano.Db.Mock.UnifiedApi
                         Test.Cardano.Db.Mock.Validate
